% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_delay.R
\name{api_delay}
\alias{api_delay}
\title{Function that controls the delay between API calls to opendota, user can specify the wait_time, but opendota asks
you to limit yourself to 1 call per second}
\usage{
api_delay(start_time, wait_time = 1)
}
\arguments{
\item{start_time}{Time of last API call}

\item{wait_time}{Desired wait time between API calls}
}
\value{
There is no return, it simply sleeps the system for whatever time needed to reach wait_time
}
\description{
Function that controls the delay between API calls to opendota, user can specify the wait_time, but opendota asks
you to limit yourself to 1 call per second
}
\examples{
\dontrun{
api_delay(start_time, wait_time)
}
}
